Ensure component focused tests live in colocated similarly named files
