+++ src/renderer/src/features/gcode-options/components/OutputSection.test.tsx
@@ -1,0 +1,31 @@
+import { describe, it, expect, vi } from "vitest";
+import { render, screen, fireEvent } from "@testing-library/react";
+import { OutputSection } from "./OutputSection";
+import { DEFAULT_GCODE_PREFS } from "../gcodePrefs";
+
+describe("OutputSection", () => {
+  it("shows connection/group warnings and forwards toggles", () => {
+    const onToggleOpen = vi.fn();
+    const onTogglePref = vi.fn();
+
+    render(
+      <OutputSection
+        open
+        connected={false}
+        layerGroupCount={0}
+        prefs={{ ...DEFAULT_GCODE_PREFS, exportPerGroup: true }}
+        onToggleOpen={onToggleOpen}
+        onTogglePref={onTogglePref}
+      />,
+    );
+
+    expect(screen.getByText("(not connected — will be skipped)")).toBeDefined();
+    expect(screen.getByText(/No groups defined/i)).toBeDefined();
+
+    fireEvent.click(screen.getByRole("button", { name: "Output" }));
+    expect(onToggleOpen).toHaveBeenCalledTimes(1);
+
+    fireEvent.click(screen.getByRole("checkbox", { name: "Save to computer" }));
+    expect(onTogglePref).toHaveBeenCalledWith("saveLocally");
+  });
+});
