+++ src/renderer/src/features/properties-panel/components/ToolpathSection.test.tsx
@@ -1,9 +1,7 @@
 import { describe, it, expect, vi } from "vitest";
 import { render, screen, fireEvent } from "@testing-library/react";
-import { ToolpathSection } from "../../src/renderer/src/features/properties-panel/components/ToolpathSection";
-import { LayersHeader } from "../../src/renderer/src/features/properties-panel/components/LayersHeader";
-import { NumberField } from "../../src/renderer/src/features/properties-panel/components/NumberField";
-import type { GcodeToolpath } from "../../src/renderer/src/utils/gcodeParser";
+import { ToolpathSection } from "./ToolpathSection";
+import type { GcodeToolpath } from "../../../utils/gcodeParser";
 
 function buildToolpath(patch: Partial<GcodeToolpath> = {}): GcodeToolpath {
   return {
@@ -19,8 +17,8 @@ function buildToolpath(patch: Partial<GcodeToolpath> = {}): GcodeToolpath {
   };
 }
 
-describe("Properties panel extracted components", () => {
-  it("ToolpathSection forwards toggle and clear actions", () => {
+describe("ToolpathSection", () => {
+  it("forwards toggle and clear actions", () => {
     const onToggleSelected = vi.fn();
     const onClear = vi.fn();
 
@@ -50,7 +48,7 @@ describe("Properties panel extracted components", () => {
     expect(screen.getByText("Est. duration")).toBeDefined();
   });
 
-  it("ToolpathSection hides feedrate row when feedrate is zero", () => {
+  it("hides feedrate row when feedrate is zero", () => {
     render(
       <ToolpathSection
         toolpath={buildToolpath({ feedrate: 0 })}
@@ -64,38 +62,5 @@ describe("Properties panel extracted components", () => {
     );
 
     expect(screen.queryByText("Feedrate")).toBeNull();
-  });
-
-  it("LayersHeader renders only when show=true and triggers add action", () => {
-    const onAddGroup = vi.fn();
-    const { rerender } = render(
-      <LayersHeader show={false} onAddGroup={onAddGroup} />,
-    );
-
-    expect(screen.queryByText("Layers")).toBeNull();
-
-    rerender(<LayersHeader show onAddGroup={onAddGroup} />);
-    expect(screen.getByText("Layers")).toBeDefined();
-
-    fireEvent.click(screen.getByRole("button", { name: "+" }));
-    expect(onAddGroup).toHaveBeenCalledTimes(1);
-  });
-
-  it("NumberField renders and emits parsed numeric changes", () => {
-    const onChange = vi.fn();
-
-    render(
-      <NumberField
-        label="X (mm)"
-        value={12.5}
-        onChange={onChange}
-        step={0.5}
-      />,
-    );
-
-    const input = screen.getByLabelText("X (mm)");
-    fireEvent.change(input, { target: { value: "15.25" } });
-
-    expect(onChange).toHaveBeenCalledWith(15.25);
   });
 });
